+++ fastagency/api/openapi/security.py
@@ -47,6 +47,9 @@ class BaseSecurity(BaseModel):
             if sub_class.is_supported(type, in_value):
                 return sub_class.__name__
         else:
+            logger.error(
+                f"Unsupported type '{type}' and in_value '{in_value}' combination"
+            )
             raise ValueError(
                 f"Unsupported type '{type}' and in_value '{in_value}' combination"
             )
